refactor: optimize code
